requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
cached-property==1.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
djangorestframework==3.13.1
idna==3.3
isodate==0.6.1
lxml==4.9.0
platformdirs==2.5.2
pytz==2022.1
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
zeep==4.1.0
